Move supplier business logic to operations folder
